Fix a hanging issue when processing empty data. (#841)
* Fix a hanging issue when processing empty data.

The output iterator will wait on the batch queue when calling `hasNext`,
and suppose to be waked up when the Python runner inserts something into
the batch queue. But the insertion will never happen if the input data
is empty. So it hangs forever.

The solution is to let the Python runner always wake up the output iterator
after it finishes the data writing by calling the new added API `finish()`.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>

* Add tests for processing empty data.

The 'small_data' is small enough to let some tasks get
no data when running.

Now only test this for the Scalar type who just implements
the columnar pipeline.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>